+++ core/mod.ts
@@ -3,7 +3,7 @@ export * from "./authorization.ts";
 export * from "./configuration.ts";
 export * from "./fetch-router.ts";
 export * from "./http.ts";
-export * from "./jwt.ts";
+export * from "./tokens.ts";
 export * from "./migrator.ts";
 export * from "./postgres.ts";
 export * from "./random.ts";